Move common test code under api directory
